runtime: tensor_gpu_run_msl int params i64 (AOT codegen emits i64 for Int args)
tensor_gpu_run_msl used `usize` for out_len/threads, but the rss AOT backend lowers
Int arguments to i64 literals (e.g. `4i64`), causing an E0308 type mismatch in the
generated Rust whenever a host program calls Metal.gpu_run_msl under AOT (it compiled
only under the reg-VM, which casts at the boundary). Match the runtime-ABI convention
used by every other Int-param intrinsic (i64) and narrow to usize at the Metal FFI
boundary. reg_vm pass-through updated to i64 accordingly.

+++ crates/rsscript/src/reg_vm/mod.rs
@@ -9714,8 +9714,8 @@ impl RegVm {
                     .iter()
                     .map(expect_float_list_ref)
                     .collect::<Result<_, _>>()?;
-                let out_len = expect_int_ref(intrinsic_arg(&self.stack, base, args, 3)?)? as usize;
-                let threads = expect_int_ref(intrinsic_arg(&self.stack, base, args, 4)?)? as usize;
+                let out_len = expect_int_ref(intrinsic_arg(&self.stack, base, args, 3)?)?;
+                let threads = expect_int_ref(intrinsic_arg(&self.stack, base, args, 4)?)?;
                 Ok(json_result(
                     match rsscript_runtime::tensor_gpu_run_msl(
                         &source, &fn_name, &inputs, out_len, threads,
